Rename uts46bis to uts46
