+++ idna/tests/tests.rs
@@ -2,7 +2,7 @@ use tester as test;
 
 mod punycode;
 mod deprecated;
-mod uts46bis;
+mod uts46;
 
 fn main() {
     let mut tests = Vec::new();
@@ -21,7 +21,7 @@ fn main() {
         };
         punycode::collect_tests(&mut add_test);
         deprecated::collect_tests(&mut add_test);
-        uts46bis::collect_tests(&mut add_test);
+        uts46::collect_tests(&mut add_test);
     }
     test::test_main(&std::env::args().collect::<Vec<_>>(), tests, None)
 }
